Move `BinBuilder` to own module
